feat: move budget year selector onto the person-selector row
PersonLayout gains an optional actions slot rendered inline with the
person picker. BudgetPage owns the year state and passes the Year select
as that action, so the year now sits on the same level as the person
instead of next to the "Budget <name>" heading below.
