+++ workday-application/src/main/react/features/budget/BudgetFeature.tsx
@@ -1,12 +1,4 @@
-import {
-  Alert,
-  FormControl,
-  InputLabel,
-  MenuItem,
-  Select,
-  Stack,
-  Typography,
-} from '@mui/material';
+import { Alert, Stack, Typography } from '@mui/material';
 import { useCallback, useEffect, useState } from 'react';
 import { BudgetClient } from '../../clients/BudgetClient';
 import type { Person } from '../../clients/PersonClient';
@@ -16,16 +8,13 @@ import { EventDialog } from '../event/EventDialog';
 import { BudgetEventsTable } from './BudgetEventsTable';
 import { BudgetSummaryCards } from './BudgetSummaryCards';
 
-const currentYear = new Date().getFullYear();
-const selectableYears = [currentYear, currentYear - 1, currentYear - 2];
-
 type BudgetFeatureProps = {
   person: Person;
+  year: number;
 };
 
-export function BudgetFeature({ person }: BudgetFeatureProps) {
+export function BudgetFeature({ person, year }: BudgetFeatureProps) {
   const [user] = useUserMe();
-  const [year, setYear] = useState(currentYear);
   const [summary, setSummary] = useState<BudgetSummaryResponse | null>(null);
   const [error, setError] = useState(false);
   const [editCode, setEditCode] = useState<string | undefined>(undefined);
@@ -53,31 +42,9 @@ export function BudgetFeature({ person }: BudgetFeatureProps) {
 
   return (
     <Stack spacing={2}>
-      <Stack
-        direction="row"
-        justifyContent="space-between"
-        alignItems="center"
-        flexWrap="wrap"
-      >
-        <Typography variant="h6">
-          Budget {person.firstname} {person.lastname}
-        </Typography>
-        <FormControl size="small" sx={{ minWidth: 120 }}>
-          <InputLabel id="budget-year-label">Year</InputLabel>
-          <Select
-            labelId="budget-year-label"
-            label="Year"
-            value={year}
-            onChange={(event) => setYear(Number(event.target.value))}
-          >
-            {selectableYears.map((selectableYear) => (
-              <MenuItem key={selectableYear} value={selectableYear}>
-                {selectableYear}
-              </MenuItem>
-            ))}
-          </Select>
-        </FormControl>
-      </Stack>
+      <Typography variant="h6">
+        Budget {person.firstname} {person.lastname}
+      </Typography>
       {error ? (
         <Alert severity="error">Could not load the budget summary.</Alert>
       ) : (
